requirements.txt
langchain_community
langchain_ibm
sentence-transformers
ipykernel
ibm-watsonx-ai
langgraph
elasticsearch
langchain-elasticsearch
elastic-transport
python-dotenv
tavily-python
langchain
#langchain_huggingface
flask
flask_cors
chromadb
langchain-chroma
beautifulsoup4
langchain_openai